feat(all): 实现完整的Sync API同步功能和数据监控
新增功能:
- 完整的Sync API实现，支持主要数据同步和种子连接监控
- 实时数据监控功能，支持增量更新和变化回调
- 新增MainData、TorrentPeer等同步数据模型
- 提供同步请求和响应对象的完整封装
- 在UnifiedClient中添加简化的同步操作方法

改进:
- 修复SearchResult和UserInfo模型中的类型错误
- 增强TransportInterface的认证管理功能
- 更新开发文档，记录Sync API完成状态

测试覆盖:
- 为所有Sync相关功能添加完整的单元测试
- 新增同步功能的集成测试和使用示例

+++ src/Model/UserInfo.php
@@ -402,7 +402,7 @@ class UserInfo implements JsonSerializable
             return null;
         }
 
-        return floor((time() - $this->createdTime) / 86400);
+        return (int) floor((time() - $this->createdTime) / 86400);
     }
 
     /**
